sih project/requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
numpy
requests 
sentence-transformers
pypdf